Fixing numerical derivation and integration

+++ sources/NumericalSolver/num_derivation.cpp
@@ -2,5 +2,5 @@
 
 
 double derivative(double f(double), double a){
-    return (int)((f(a+1e-10)-f(a-1e-10))/(2e-10) * 100000.0)/100000.0;
+    return ((f(a+1e-10)-f(a-1e-10))/(2e-10) * 100000.0)/100000.0;
 };
